maj package

+++ build_push_package.sh
@@ -40,22 +40,29 @@ if [[ -n $1 ]]; then
     echo "ERROR : $1 mauvaise option"
     exit 1
   fi
+
+  if [[ -n $2 ]]; then
+    # deploie en local via git
+    if [[ $1 = "--deploy" ]]; then
+        echo "---- DEPLOY ........................................................."
+        pip3 uninstall -y terms_tools
+        pip3 install --no-cache-dir git+http://vxgit.intra.inist.fr:60000/git/schneist/terms_tools.git@${tag}#egg=terms_tools
+
+        # info se mettre en contexte execution avec le paquet installé
+        echo "---- CONTROL DEPLOY ........................................................."
+        pip3 show -v terms-tools
+        unset PYTHONPATH
+
+        echo "---- EXECUTION TEST (lib mode) ........................................................."
+        cmd="cat test/data/not-fr.tsv| terms_tools POStagger -f text -o doc -log analyze.log -lang fr"
+        eval $cmd
+    fi
+
 else
   git push http://schneist:merlin@vxgit.intra.inist.fr:60000/git/schneist/terms_tools.git --all
 fi
 
-# deploie en local via git
-echo "---- DEPLOY ........................................................."
-pip3 uninstall -y terms_tools
-pip3 install --no-cache-dir git+http://vxgit.intra.inist.fr:60000/git/schneist/terms_tools.git@${tag}#egg=terms_tools
 
-# info se mettre en contexte execution avec le paquet installé
-echo "---- CONTROL DEPLOY ........................................................."
-pip3 show -v terms-tools
-unset PYTHONPATH
 
-echo "---- EXECUTION TEST (lib mode) ........................................................."
-cmd="cat test/data/not-fr.tsv| terms_tools POStagger -f text -o doc -log analyze.log -lang fr"
-eval $cmd
